# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WhyQuiet API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/batches": {
      "get": {
        "summary": "List Batches",
        "operationId": "list_batches_api_batches_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BatchStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Batch"
                  },
                  "title": "Response List Batches Api Batches Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Batch",
        "operationId": "create_batch_api_batches_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Batch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/batches/{id}/approve": {
      "post": {
        "summary": "Approve Batch",
        "operationId": "approve_batch_api_batches__id__approve_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Batch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/batches/{id}/reject": {
      "post": {
        "summary": "Reject Batch",
        "operationId": "reject_batch_api_batches__id__reject_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Batch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/batches/{id}/export": {
      "get": {
        "summary": "Export Batch",
        "operationId": "export_batch_api_batches__id__export_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "json",
                "csv"
              ],
              "type": "string",
              "default": "json",
              "title": "Format"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportBatchResponse"
                }
              },
              "text/csv": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/campaign/receipts": {
      "post": {
        "summary": "Campaign Receipt",
        "operationId": "campaign_receipt_api_campaign_receipts_post",
        "parameters": [
          {
            "name": "x-whyquiet-timestamp",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Whyquiet-Timestamp"
            }
          },
          {
            "name": "x-whyquiet-signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Whyquiet-Signature"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Receipt"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/wallets/locked": {
      "get": {
        "summary": "List Locked Wallets",
        "operationId": "list_locked_wallets_api_wallets_locked_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LockedWallet"
                  },
                  "title": "Response List Locked Wallets Api Wallets Locked Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/batches/{id}/audit": {
      "get": {
        "summary": "Get Batch Audit",
        "operationId": "get_batch_audit_api_batches__id__audit_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEntry"
                  },
                  "title": "Response Get Batch Audit Api Batches  Id  Audit Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Batch": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "cause": {
            "$ref": "#/components/schemas/CauseFamily"
          },
          "remedy_code": {
            "type": "string",
            "title": "Remedy Code"
          },
          "unit_cost_bdt": {
            "type": "number",
            "title": "Unit Cost Bdt"
          },
          "wallet_count": {
            "type": "integer",
            "title": "Wallet Count"
          },
          "status": {
            "$ref": "#/components/schemas/BatchStatus"
          },
          "proposed_by": {
            "type": "string",
            "title": "Proposed By"
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decision_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Note"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "sms": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SmsStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "cause",
          "remedy_code",
          "unit_cost_bdt",
          "wallet_count",
          "status",
          "proposed_by",
          "created_at"
        ],
        "title": "Batch"
      },
      "BatchStatus": {
        "type": "string",
        "enum": [
          "proposed",
          "approved",
          "rejected"
        ],
        "title": "BatchStatus"
      },
...
      "CauseFamily": {
        "type": "string",
        "enum": [
          "job_exit",
          "migration",
          "solved_problem",
          "fee_shock",
          "supply_failure"
        ],
        "title": "CauseFamily"
      },
...
      "CreateBatchRequest": {
        "properties": {
          "cause": {
            "$ref": "#/components/schemas/CauseFamily"
          },
          "wallet_ids": {
            "items": {
              "type": "string",
              "pattern": "^W-[0-9A-Z]{6}$"
            },
            "type": "array",
            "maxItems": 1000,
            "minItems": 1,
            "title": "Wallet Ids"
          }
        },
        "type": "object",
        "required": [
          "cause",
          "wallet_ids"
        ],
        "title": "CreateBatchRequest"
      },
      "DecisionRequest": {
        "properties": {
          "note": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "note"
        ],
        "title": "DecisionRequest"
      },
...
      "ExportBatchResponse": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "cause": {
            "$ref": "#/components/schemas/CauseFamily"
          },
          "remedy_code": {
            "type": "string",
            "title": "Remedy Code"
          },
          "wallet_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Wallet Ids"
          },
          "cost_bdt": {
            "type": "number",
            "title": "Cost Bdt"
          },
          "approved_by": {
            "type": "string",
            "title": "Approved By"
          },
          "approved_at": {
            "type": "string",
            "title": "Approved At"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "cause",
          "remedy_code",
          "wallet_ids",
          "cost_bdt",
          "approved_by",
          "approved_at"
        ],
        "title": "ExportBatchResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LockReason": {
        "type": "string",
        "enum": [
          "open",
          "cooldown"
        ],
        "title": "LockReason"
      },
      "LockedWallet": {
        "properties": {
          "wallet_id": {
            "type": "string",
            "title": "Wallet Id"
          },
          "reason": {
            "$ref": "#/components/schemas/LockReason"
          },
          "until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Until"
          }
        },
        "type": "object",
        "required": [
          "wallet_id",
          "reason"
        ],
        "title": "LockedWallet"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user_id",
          "role"
        ],
        "title": "LoginResponse"
      },
...
      "Receipt": {
        "properties": {
          "batch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Batch Id"
          },
          "sent": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Sent"
          },
          "delivered": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Delivered"
          },
          "failed": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Failed"
          },
          "gateway_ref": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Gateway Ref"
          },
          "failed_wallet_ids": {
            "items": {
              "type": "string",
              "pattern": "^W-[0-9A-Z]{6}$"
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Failed Wallet Ids"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "sent",
          "delivered",
          "failed"
        ],
        "title": "Receipt"
      },
...
      "UserRole": {
        "type": "string",
        "enum": [
          "analyst",
          "approver"
        ],
        "title": "UserRole"
      },
...
    }
  }
}